Atualiza biblioteca utils e adiciona arquivo com numeros aleatorios

+++ Algoritmos-de-Ordenacao/utils.c
@@ -4,30 +4,30 @@
 #include "utils.h"
 
 
-void preenche_vetor(int vetor[], int tamanho){
+void preenche_vetor(long int vetor[], long int tamanho){
 	srand(time(NULL));
 
-	for(int i=0; i<tamanho; i++){
+	for(long int i=0; i<tamanho; i++){
 		vetor[i] = rand() % 1001;
 	}
 }
 
-void imprime_vetor(int vetor[], int tamanho){
+void imprime_vetor(long int vetor[], long int tamanho){
 	printf("\n[");
-	for(int i=0; i<tamanho-1; i++){
-		printf("%d, ", vetor[i]);
+	for(long int i=0; i<tamanho-1; i++){
+		printf("%ld, ", vetor[i]);
 	}
-	printf("%d]\n", vetor[tamanho-1]);
+	printf("%ld]\n", vetor[tamanho-1]);
 }
 
-void le_vetor(int vetor[], int tamanho){
+void le_vetor(long int vetor[], long int tamanho){
 	for(int i=0; i<tamanho; i++){
-		scanf("%d", &vetor[i]);
+		scanf("%ld", &vetor[i]);
 	}
 }
 
-void troca(int vetor[], int i, int j){
-	int temp = vetor[i];
+void troca(long int vetor[], long int i, long int j){
+	long int temp = vetor[i];
 	vetor[i] = vetor[j];
 	vetor[j] = temp;
 }
